move event content to page, install and init atproto client
