+++ package.json
@@ -11,6 +11,7 @@
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch"
 	},
 	"dependencies": {
+		"@atproto/oauth-client-browser": "^0.3.7",
 		"bits-ui": "^0.22.0",
 		"svelte-boring-avatars": "^1.2.6"
 	},
