+++ src/routes/client_metadata.json/+server.ts
@@ -1,0 +1,6 @@
+import { json } from "@sveltejs/kit";
+import { clientMetadata } from "$lib/atproto";
+
+export async function GET() {
+  return json(clientMetadata);
+}
